impr: new imports

+++ companions/proto/proto.go
@@ -111,12 +111,13 @@ func (g *Buf) Generate(ctx context.Context) error {
 	// Move the result
 	openapi := path.Join(g.Dir, "openapi/api.swagger.json")
 	destination := path.Join(g.Dir, standards.OpenAPIPath)
-	err = shared.CopyFile(ctx, openapi, destination)
-	if err != nil {
-		return w.Wrapf(err, "cannot copy file")
+	if openapi != destination {
+		err = shared.CopyFile(ctx, openapi, destination)
+		if err != nil {
+			return w.Wrapf(err, "cannot copy file")
+		}
+		_ = os.Remove(openapi)
 	}
-	_ = os.Remove(openapi)
-
 	err = g.dependencies.UpdateCache(ctx)
 	if err != nil {
 		return w.Wrapf(err, "cannot update cache")
